refactor: delete execution_engine_anonymizer
